requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langgraph
supabase
resend
python-dotenv
pydantic
schedule
google-genai
Pillow
streamlit
python-multipart
